spaces sidebar / starting new notes page
