+++ apps/frontend/src/app/app/space/page.tsx
@@ -1,13 +1,11 @@
 import * as React from "react"
 
-import Container from "@/src/components/atoms/Container"
-
 const SpacePage: React.FC = () => {
   return (
-    <section className="h-full overflow-auto rounded-xl border border-white/10 bg-white/10 px-6 py-16 shadow-lg backdrop-blur-lg">
-      <Container>
-        <p className="text-zinc-400">Select a space from the sidebar</p>
-      </Container>
+    <section className="w-full h-full overflow-auto px-8 py-16 bg-secondary">
+      <p className="text-secondary-foreground">
+        Select a space from the sidebar
+      </p>
     </section>
   )
 }
